[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.03 ms, Memory: 73.6 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,7 +1,7 @@
 class Solution {
     public long[] solution(int x, int n) {
         long[] arr = new long[n];
-        arr[0] = (long) x;
+        arr[0] = x;
         
         for(int i = 0; i < n; i++){
             arr[i] = (long)x * (i+1);
